+++ playwright.config.ts
@@ -13,6 +13,10 @@ import { defineConfig, devices } from "@playwright/test";
  */
 export default defineConfig({
   testDir: "./playwright",
+  timeout: 30000, // default timeout for each test
+  expect: {
+    timeout: 5000, // default timeout for expect statements
+  },
   /* Run tests in files in parallel */
   fullyParallel: true,
   /* Fail the build on CI if you accidentally left test.only in the source code. */
@@ -70,14 +74,10 @@ export default defineConfig({
     // },
   ],
 
-    /* Run your local dev server before starting the tests */
-  webServer: !!process.env.CI
-    ? undefined
-    : {
-        // only run in local environment not CI
-        command: "npm run dev",
-        url: "http://localhost:3000/library-card/new",
-        reuseExistingServer: !process.env.CI,
-      },
+  /* Run your local dev server before starting the tests */
+  webServer: {
+    command: "npm run dev",
+    url: "http://localhost:3000/library-card/new",
+    reuseExistingServer: !process.env.CI,
+  },
 });
-